rename fiels a move the
